Add fallback support for the Secret portal.
This allows flatpaks (and whoever else) to use the Secrets portal -
these apps can then use it to aid things such as encryption.

The gnome-keyring package provides this capability, but its
portal definition is restricted to gnome sessions only. The file
being added here will allow it to work in Xfce4 and MATE as well.

Cinnamon doesn't really need this as it provides a new, preferred
portal definition file which already has gnome-keyring enabled.

ref:
https://github.com/linuxmint/cinnamon/blob/master/data/xdg-portal/x-cinnamon-portals.conf

Fixes #19.