Don't try to initialize gtk, use GMainLoop instead.
The xdg-desktop-portal process is not exiting when the session
ends (because of linuxmint/cinnamon-session@a5d4be9515 - the
session bus is never killed).

Because its background implementation regularly polls for app
states, it can trigger xdg-desktop-portal-xapp to launch while
no session is running, and no xserver (for that dead session).

This fails because gtk_init() kills our xapp portal process
immediately, and xdg-desktop-portal is stuck waiting on a dbus
response that will never arrive. It eventually times out, and
things generally right themselves.

The Gtk code was added back while refactoring the settings portal.
We've ended up not using any gtk code for now, so this can just
be removed.

If we ever do need to ignore gtk failure in the future (if we
don't fix the cinnamon-session bug), we might use gtk_init_check
instead of gtk_init. This will fail but not terminate the program.

The main issue here is that xdg-desktop-portal isn't terminating,
and/or shouldn't be requesting app states when there's no active
session. This needs to be fixed in cinnamon-session.

Fixes: #20
Fixes: linuxmint/cinnamon#12321